requirements/base.txt
torch
transformers
datasets
openai
yt-dlp
requests
google-generativeai
